[flutter_tools] rename library to be less absurd
