+++ packages/flutter_tools/lib/src/aot.dart
@@ -10,7 +10,7 @@ import 'base/common.dart';
 import 'base/logger.dart';
 import 'build_info.dart';
 import 'build_system/build_system.dart';
-import 'build_system/targets/dart.dart';
+import 'build_system/targets/common.dart';
 import 'build_system/targets/icon_tree_shaker.dart';
 import 'build_system/targets/ios.dart';
 import 'cache.dart';
