+++ packages/flutter_tools/lib/src/build_system/targets/macos.dart
@@ -13,7 +13,7 @@ import '../build_system.dart';
 import '../depfile.dart';
 import '../exceptions.dart';
 import 'assets.dart';
-import 'dart.dart';
+import 'common.dart';
 import 'icon_tree_shaker.dart';
 
 /// Copy the macOS framework to the correct copy dir by invoking 'cp -R'.
