+++ packages/flutter_tools/lib/src/bundle.dart
@@ -14,7 +14,7 @@ import 'base/logger.dart';
 import 'build_info.dart';
 import 'build_system/build_system.dart';
 import 'build_system/depfile.dart';
-import 'build_system/targets/dart.dart';
+import 'build_system/targets/common.dart';
 import 'build_system/targets/icon_tree_shaker.dart';
 import 'cache.dart';
 import 'dart/package_map.dart';
